refactor delete endpoint to accept sys_id

+++ src/main.py
@@ -263,8 +263,8 @@ def sync_device(device_body: DeviceBody, background_tasks: BackgroundTasks):
 
 @logger.catch
 @app.delete('/devices/{id}', dependencies=[Depends(authorize)])
-def delete_device(prtg_url: str, prtg_api_key: str, background_tasks: BackgroundTasks, prtg_id: str = Path(..., alias='id'), request_id: str | None = None):
-    background_tasks.add_task(delete_device_task, prtg_id, prtg_url, prtg_api_key, request_id)
+def delete_device(prtg_url: str, prtg_api_key: str, background_tasks: BackgroundTasks, sys_id: str = Path(..., alias='id'), request_id: str | None = None):
+    background_tasks.add_task(delete_device_task, sys_id, prtg_url, prtg_api_key, request_id)
 
 
 def log_error_console_and_snow(request_id: str | None, error_msg: str):
@@ -417,24 +417,29 @@ def sync_device_task(device_body: DeviceBody):
         success_log = Log(device_body.request_id, State.SUCCESS, f'Successfully created/updated {device.name} with ID {device.id}.')
         snow_controller.post_log(success_log)
 
-def delete_device_task(prtg_id: str, prtg_url: str, prtg_api_key: str, request_id: str | None):
+def delete_device_task(sys_id: str, prtg_url: str, prtg_api_key: str, request_id: str | None):
     # build PRTG controller from parameters
     auth = BasicToken(prtg_api_key)
     client = PrtgClient(prtg_url, auth)
     prtg_controller = PrtgController(client)
     logger.debug(f'PRTG URL: {prtg_url}')
-    logger.debug(f'Device ID from payload: {id}.')
+    logger.debug(f'Device ID from payload: {sys_id}.')
 
     try:
-        prtg_device = prtg_controller.get_device(prtg_id)
-    except ValueError as e:
+        ci = snow_controller.get_config_item(sys_id)
+    except (ValueError, TypeError) as e:
         log_error_console_and_snow(request_id, str(e))
         return
+    if ci.prtg_id is None:
+        raise HTTPException(status.HTTP_404_NOT_FOUND, f'Cannot find PRTG ID for device {ci.name}.')
     try:
+        prtg_device = prtg_controller.get_device(ci.prtg_id)
         prtg_controller.delete_object(prtg_device)
     except ValueError as e:
         log_error_console_and_snow(request_id, str(e))
         return
+    ci.prtg_id = None
+    snow_controller.update_config_item(ci)
     msg = f'Successfully deleted {prtg_device.name} with ID {prtg_device.id}.'
     logger.info(msg)
     if request_id is not None:
